Move prop type definitions into external file for export purposes.
