+++ src/index.ts
@@ -2,7 +2,6 @@ import { BackgroundImageset, skyBackgroundImagesets } from "./background";
 import { useFullscreen, useWindowShape, WindowShape } from "./composables";
 import { PositionCoords, useGeolocation } from "./geolocation";
 import { KeyboardControlSettings } from "./keyboard";
-import { API_BASE_URL, MINIDS_BASE_URL, R2D, D2R, supportsTouchscreen, isMobile, blurActiveElement, filterInPlace } from "./utils";
 import { useWWTKeyboardControls } from "./wwt-utils";
 
 import IconButton from "./components/IconButton.vue";
@@ -12,6 +11,7 @@ import CreditLogos from "./components/CreditLogos.vue";
 import GeolocationButton from "./components/GeolocationButton.vue";
 import wwtHUD from "./components/wwtHUD.vue";
 import Gallery from "./components/Gallery.vue";
+
 
 export {
   BackgroundImageset,
@@ -30,12 +30,7 @@ export {
   PositionCoords,
   useGeolocation,
   useWWTKeyboardControls,
-  supportsTouchscreen,
-  isMobile,
-  blurActiveElement,
-  filterInPlace,
-  API_BASE_URL,
-  MINIDS_BASE_URL,
-  R2D,
-  D2R
 };
+
+export * from "./types";
+export * from "./utils";
